chore(go): remove ruleset linter package.
