+++ pkg/bundle/ruleset/engine/cel/engine.go
@@ -28,8 +28,8 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	bundlev1 "github.com/elastic/harp/api/gen/go/harp/bundle/v1"
-	"github.com/elastic/harp/pkg/bundle/ruleset/linter/engine"
-	"github.com/elastic/harp/pkg/bundle/ruleset/linter/engine/cel/ext"
+	"github.com/elastic/harp/pkg/bundle/ruleset/engine"
+	"github.com/elastic/harp/pkg/bundle/ruleset/engine/cel/ext"
 )
 
 // -----------------------------------------------------------------------------
