+++ pkg/bundle/ruleset/engine/cel/engine_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	bundlev1 "github.com/elastic/harp/api/gen/go/harp/bundle/v1"
-	"github.com/elastic/harp/pkg/bundle/ruleset/linter/engine"
+	"github.com/elastic/harp/pkg/bundle/ruleset/engine"
 	"github.com/elastic/harp/pkg/bundle/secret"
 )
