+++ pkg/bundle/ruleset/reader.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package linter
+package ruleset
 
 import (
 	"fmt"
